+++ HW_210521_ModuleTutorial/demo/src/main/java/com.metanit.hello/Hello.java
@@ -6,8 +6,6 @@ import com.metanit.factorial.Factorial;
 
 public class Hello {
     public static void main(String[] args) {
-        //System.out.println("Hello Demo Module!");
-        //JOptionPane.showMessageDialog(null, "Hello Demo Module!");
         for (int i = 0; i < 11; i++){
             int x = Factorial.calculate(i);
             String text = "Factorial of " + i + " is equal to " + x + ".";
